Add numeric comparison operators for FPI values
These are defined, so far, to mean:

 - eq: a == b if there's any overlap between a and b
 - neq: a != b if there's any of a not in b, or any of b not in a
 - lt: a < b if a is provably smaller than b (no overlap)
 - lte: a <= b if is smaller than or overlaps b
 - gt: a > b if a is provably greater than b (no overlap)
 - gte: a >= b if is greater than or overlaps b

Note that:

 - (a == b) does not imply (a != b)
 - (a >= b) does not imply !(a <= b)
 - etc.

This still requires further work.  First, are these definitions useful?
Second, can they be relied upon given that we're in an abstract over a
core type and while doing so significantly changing the semantics
of the basic comparisons?  Third, are they coherent?

Finally, this needs to be thoroughly documented.